flaskbackend/requirements.txt
Flask
psycopg2
flask-cors
opentelemetry-api
opentelemetry-sdk 
opentelemetry-instrumentation-flask